requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
djangorestframework==3.13.1
drf-yasg==1.20.0
html2text==2020.1.16
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.0
nltk==3.7
packaging==21.3
pyparsing==3.0.7
pytz==2021.3
regex==2022.3.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tqdm==4.63.0
uritemplate==4.1.1
urllib3==1.26.8
